requirements.txt
# Requirements for browser-calls-django

# Django (1.8 is a long-term support release)
Django==1.8

# Third-party app dependencies
django-bootstrap3==6.2.2
django-forms-bootstrap==3.0.1
django-phonenumber-field==0.7.2
psycopg2==2.6.1
whitenoise==2.0.4

# Production dependencies
gunicorn==19.3.0

# Test dependencies
coverage==4.0
mock==1.3.0
model-mommy==1.2.5
six==1.9.0
